rts-worker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rtsworker"]

[project]
name = "rts-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.10.5",
    "pyserial>=3.5",
    "requests>=2.32.3",
]

[project.scripts]
rtsworker = "main:main"